refactor(pcc-service-pack-options): rename fonctions to uniformize
